move structs to own file
